XRENDERING-736: Macros info, success, warning and error are only distinguished by colors * Moved the IconProvider classes to util instead of transformation
